Refactor webpack build chain (add dist files), fix reset time event error, resolves #27
