+++ css/appointment-picker.scss
@@ -1,5 +1,5 @@
 ﻿/*
- * appointment-picker.css 1.2.0 | MIT License | github.com/jannicz/appointment-picker
+ * appointment-picker.css 1.3.0 | MIT License | github.com/jannicz/appointment-picker
  */
 
 /* Default variation */
