--- example-react/react-embed.html
@@ -1,11 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-	<meta charset="UTF-8">
-	<title>React Integration Example</title>
-</head>
-<body>
-	<div id="app"></div>
-	<script src="./bundle.js"></script>
-</body>
-</html>
