+++ example/react.html
@@ -1,0 +1,15 @@
+<!DOCTYPE html>
+<html>
+<head>
+	<meta charset="UTF-8">
+	<title>React Integration Example</title>
+</head>
+<body>
+	<article>
+		<div id="app"></div>
+		<p>
+			<a href="https://github.com/jannicz/appointment-picker/tree/master/example/react/react-component.js">React Component source</a>
+		</p>
+	</article>
+<script type="text/javascript" src="react-component.min.js"></script></body>
+</html>
