+++ index.html
@@ -125,7 +125,7 @@ new AppointmentPicker(document.getElementById('time-4'), {
 	<p>
 		<h3>More Examples</h3>
 		<ul>
-			<li><a href="example-react/react-embed.html">React Integration Example</a></li>
+			<li><a href="example/react.html">React Integration Example</a></li>
 			<li><a href="example/exposed-functions.html">Calling exposed functions</a></li>
 			<li><a href="example/form-submit.html">Form embed Example</a></li>
 			<li><a href="example/render-on-init.html">Rendering into DOM on init</a></li>
