+++ package.json
@@ -2,11 +2,12 @@
   "name": "appointment-picker",
   "version": "1.3.0",
   "description": "customizable timepicker to be used with or without jQuery",
-  "main": "js/appointment-picker.js",
+  "main": "js/appointment-picker.min.js",
   "scripts": {
     "uglify": "uglifyjs ./js/appointment-picker.js -c -m -o ./js/appointment-picker.min.js",
     "css": "node-sass css/appointment-picker.scss css/appointment-picker.css",
-    "start": "webpack-dev-server --config ./webpack.config.js --mode development"
+    "start": "webpack-dev-server --config ./webpack.config.js --mode development --open",
+    "build": "npm run uglify && npm run css && webpack -p --env=prod --config ./webpack.prod.config.js"
   },
   "repository": {
     "type": "git",
